Applied HARMONY-3236 [drlvm][winx64][build] Fixed majority of warnings.
I've had to change Runtime_win.cpp to make it compile without warnings on win32.
I had to add include <stddef.h> to dump.h to make it compile on linux.

Tests passed on Ubuntu6 x86, Windows XP x86, SuSE9 x86_64 and Gentoo x86_64.
On windows x86_64 compilation passes.



git-svn-id: https://svn.apache.org/repos/asf/harmony/enhanced/drlvm/trunk@510652 13f79535-47bb-0310-9956-ffa450edef68

+++ vm/port/src/lil/lil_code_generator.cpp
@@ -64,7 +64,8 @@ NativeCodePtr LilCodeGenerator::compile(LilCodeStub* cs, PoolManager* code_pool)
     compile_add_dynamic_generated_code_chunk("unknown", stub, stub_size);
 
     if (VM_Global_State::loader_env->TI->isEnabled())
-        jvmti_send_dynamic_code_generated_event("unknown", stub, stub_size);
+        jvmti_send_dynamic_code_generated_event("unknown", stub,
+            (jint)stub_size);
 
     return stub;
 }
